Assignment_3/flaskProject2/requirements.txt
flask 
pymongo 
dnspython 
python-dotenv